Misc Bug Fixes
- WeaponSwitch now doesn't re-instantiate weapon prefabs by repeatedly pressing the equip key
- Fixed TakeDamage function by returning it to an int function

+++ Assets/Scripts/EnemyAI/EnemyAI.cs
@@ -4,9 +4,9 @@ public class EnemyAI : MonoBehaviour
 {
     public float moveSpeed = 3f;
     protected Transform player;
-    public float damage = 1;
-    public float health = 3;
-    public float maxHealth = 3;
+    public int damage = 1;
+    public int health = 3;
+    public int maxHealth = 3;
     public SpriteRenderer healthBar;
     private Vector2 hbScale;
